[Ronan Lehane] ECK Support Clarification
Added the following clarification regarding licensing for ECK as we are getting a significant numbers of cases from Gold/Platinum customers asking for support on ECK. ECK support is only available to Enterprise customers. 

IMPORTANT: ECK is offered in two licensing tiers only: basic and enterprise. Similar to the Elastic Stack today, customers can download and use ECK for free but a subscription is required for support. ECK support requires an Enterprise subscription.