Consistently rely on the global scheme.Scheme across the codebase (#2690)
We manipulate the scheme inconsistently across the code base.
Sometimes through the global scheme.Scheme from client-go, sometimes by
passing it as a parameter across the entire function stack.

It turns out the parameter we pass to functions is just a pointer
reference to the global scheme.Scheme, so we loose any benefit from not
using a global variable.

Therefore, I think it's safe (and less verbose), to just rely on the
global client go scheme everywhere. The only gotcha is to make sure it
is properly initialized with all CRDs, which is done when the manager
bootstraps, and when we create the fake k8s client in unit tests.