Prevent pod deletion while ES node still contains shards (#2715)
Removes the logic that would declare a node evacuation as completed as soon as there was another copy of every shard on the node to be deleted on at least one other node.

Instead we now require nodes to be completely evacuated before deletion can proceed.

This opens up the possibility for downscale operations that cannot make progress. For now we only adjust the documentation to address this and advise users to adjust their index settings.