Fix TestMutationHTTPToHTTPS and TestMutationHTTPSToHTTP test builders
Our mutation test logic relies on computing a hash of the existing Pod
and comparing it with the hash of mutated Pods.
This only works fine if builder.MutatedFrom is correctly set, which is
done in `RunESMutation()`.

Because this function was not used in the 2 tests, the hash comparison
was not correct.

This highlights how our e2e test builder framework is quite brittle.