Fix red cluster healh reported on tls toggle e2e test
When computing the number of replicas for the data integrity check, we
inspect the ES specs to figure out whether a rolling upgrade will
happen.

We trigger a rolling upgrade when TLS is toggled in the HTTP options,
but the test was checking the opposite (deepEqual instead of
!deepEqual).

As a result the data integrity index was created with 0 replicas, since
we though this was going to be a grown-and-shrink upgrade.