Close idle connections on the http client directly
I realized while refactoring this piece of code for being used by
another controller that we don't need to keep track of the
http.Transport, since calling Close() on the client propagates the close
to the underlying transport directly.

This simplifies the ES client structure a little bit, and will be useful
for the refactoring I'm working on.