Factorize HTTP certs reconciliation in a single function (#2722)
Add a single common function to deal with HTTP certificates reconciliation.
It consists in reconciling 3 secrets:

- internal ca certs
- internal http certs
- public http certs (no private key)

The immediate benefit is to remove ~80 lines of somewhat duplicated code in our 4 main controllers.

This is mostly arranging code around. However, there are a few important changes to notice:

Internal and CA secrets were created with labels matching the resource (eg. ES cluster name), but not public certs. This PR applies the same labels to all 3 secrets.
The 3 secrets are removed automatically if TLS is disabled for Kibana, APMServer and Enterprise Search. But not for Elasticsearch, which may benefit from TLS certs hot-reload in case TLS is toggled on/off/on quickly.