Request a flush (not a synced flush) when expected ES is 8.0+
Synced flush is deprecated starting 7.6, and removed starting 8.0.
Instead, we should perform a "normal" flush during rolling upgrades.

That's what this commit does.

One important implementation detail: it only inspects the expected ES
version, not the version of the currently running Pods, for simplicity.
When upgrading from 7.x to 8.x, a normal flush will be done.
In theory, we could do a synced flush for the first
pod to upgrade (since there are no 8.0 pods yet), then a normal flush
for the remaining pods (since the request may target the already
upgraded 8.0 pod).  We must make sure we don't do a synced flush if
an 8.x pod is running, or should be running.
Since synced flush is a best effort operation, and
7.x does also support normal flush, I think it's not worth the trouble.

I noticed the current 8.0.0-SNAPSHOT still accepts `/_flush/synced` requests,
but that may disappear anytime.