Clarify data migration function and unit tests (#2845)
I got confused by the comment attached to the `IsMigratingData`
function, along with its unit tests. They somehow indicate we're
checking if a node has the only copy of a shard.
But the reality is we just check if there's any shard on that node.

This commit renames the function to `ShardsOnNode()`, modifies the
comment accordingly, and adapt the unit tests to the very simple logic
of the function.