Check node excluded from shard allocation before deleting
This amends the checks we do to verify node evacuation is complete. Previously we only checked that no shards are present on the node in question.
This however was not good enough as a missing allocation filter for that node could mean that shards move back to the node just before deletion.
The new check looks at allocation filters and shards. While this check is not perfect as allocation filters can be removed before deletions it narrows
window for that to happen somewhat.