Use the health endpoint for Enterprise Search readiness probe
Recent versions include a new health endpoint, that must be requested
with an Elasticsearch user.

Even though this user technically only requires "monitor" permissions on
the ES cluster, I decided to use the same ES user as Enterprise Search
is already using. For the sake of simplicity, but also because users may
run Enterprise Searrch without an associated ES resource managed by ECK.
In which case we cannot create a custom user with monitor permissions.

The readiness probe script is stored in the same secret as the
Enterprise Search configuration, to avoid mounting an extra secret with
the user credentials. Both the configuration and the readiness probe
script are reconciled at the same time.