Workaround TestDeleteServices flakiness by not checking the endpoints (#3087)
This test fails once in a while (~1/200 run fails). It seems that we are
hitting a strange bug or race condition in either K8s or GKE, which
leads to the endpoints resource not being recreated immediately after we
delete then recreate the service. The endpoints resource eventually
appears, hours later.

We have investigated this issue for a while and are fairly confident
there is no bug in ECK that could cause that situation.
While there's still a potential issue to fix somewhere else, let's not
keep this flaky test around causing too much noise.

Instead of checking the entire state of the cluster, which includes a
check on the Endpoints resource, let's just check the service we delete
is correctly recreated.
There is value in still checking the Endpoints resource as part of every
other test, which do not involve a quick service
creation/deletion/creation.