Do not stop running trial even when ECK not in trial mode
Trials on Elasticsearch level can only be started once.
The intention behind this change is avoid a bad user experience.
We encourage users to start trials via ECK but the Kibana UI also
advertises trials in various places when users try to access commercial features.
We don't want to change the Kibana behaviour to keep this easy way to try out a feature.
This change therefore tolerates a cluster running in trial mode even if ECK's license level
would require us to downgrade this cluster to a Basic license.