Perform a sync call to retrieve ES license in the reconciliation
We used to rely on the license fetched by the observers on a regular
basis, but this can lead to a race condition as described in
https://github.com/elastic/cloud-on-k8s/issues/3163#issuecomment-635875886.

Instead, let's keep things simple, and do a synchronous request to
get the license as part of the normal reconciliation, to then decide
whether or not it should be updated.

In case of error (eg. Elasticsearch req timeout), enqueue a new
reconciliation, but don't abort the current reconciliation.