Replace the license observer by a sync call in the reconciliation (#3165)
Fixes #3163.
Fixes a bug introduced in #3150.

We used to rely on the license fetched by the observers on a regular
basis, but this can lead to a race condition as described in
#3163 (comment).

Instead, let's keep things simple, and do a synchronous request to
get the license as part of the normal reconciliation, to then decide
whether or not it should be updated.

In case of error (eg. Elasticsearch req timeout), enqueue a new
reconciliation, but don't abort the current reconciliation.

This PR also adds some minor refactoring of the code, and avoids
a call to start_basic if the current license is already basic.