Use _cluster/health API during rolling upgrades (#3195)
This is to make sure there are no pending or ongoing shard fetches,
initializations or relocations when rolling the next node. Otherwise
the cluster status might go red if we roll forward before the indices
on the previous node have recovered. This applies to upgrades
only (i.e. version upgrades)