Allow single-node clusters disruption in the default PDB
See issue PDB per Node type (Master, Data, Ingest) #2936

This PR will make sure that when someone deploys a single node cluster, the pdb does not get set to 1.

If someone deploys a single node cluster now, it can make sure the k8s cluster cannot reboot.
The default setting for single node, IMHO, should not set the pdb to 1.
This change makes sure that when it is single node, the pdb is 0. When the elastic cluster is scaled up, the pdb is calculated as "normal".