Fix Ent Search hash computation when TLS is disabled
Due to a missing check on whether TLS is enabled or not, the
reconciliation did not make progress when TLS was disabled.

This is fixing it by adding the extra check. The PR also unit tests the
entire buildConfigHash function.

I'll add a follow-up E2E test to cover Enterprise Search with TLS
disabled.