Suggest starting an ECK trial license in Enterprise Search docs
In order to start Workplace Search users will need an Elastic license. A
great way to try it is to start a trial license.
While a trial can be started by clicking a button on the UI, we want to
encourage users to start a trial at ECK level, so that:
* they also benefit from ECK licensed features
* they don't hit a bug in Enterprise Search 7.7.0 (fixed in 7.8.0) where
  the start trial call returns an error 500 (once)