Add an E2E test to cover Enterprise Search version upgrades
Let's double-check Enterprise Search version upgrades work as expected.
The test itself does not check that read-only mode was correctly toggled
by the operator, but the Pod in a newer version would fail if read-only
mode wasn't toggled anyway.
Similar to other stack tests, this one performs the upgrade from the
current version in the test context to the latest 7.x.