E2E work around Kibana 7.1.1 index creation issue (#3385)
Run test only with one Kibana node on 7.1.1 to avoid an issue where Kibana would accidentally create a .kibana index with a replica
instead of the expected alias. This leads to a test failure because we wait for green cluster status which is impossible to achieve on a single node cluster.