Retry getting the E2E tests Pod logs when the stream ends (#3389)
* Retry getting the E2E tests Pod logs if the stream ends

In some environments (OCP mostly), we don't manage to get a full 6 hours
long Pod log stream with no interruptions. More generally, network isn't
reliable so this could happen in any environments.

Let's retry the log stream apiserver request when that happens.

The side-effect is that we now get duplicate log entries. In CI we have a
process that parses the logs to turn them into formatted test results.
In order to not make that process fail, we ask Kubernetes to
provide a timestamp with each log, so we can keep track of the last timestamp.
On a retry, we can therefore ignore logs whose timestamp is lower than
the last one we found.
This isn't a perfect system: if the stream drops in between several logs
that share the same timestamp, the second half will be lost. In our case
for test results I think that's unlikely to happen.

* Remove unused variable