Report the lowest running version in the status of each resource
Report the currently running version in the `status.version` field of
each resource. In case a version upgrade is in progress for that
resource, report the lowest running version.

For Elasticsearch, we check both the Pods and the StatefulSets to report
the lowest version. A StatefulSet might report a higher version whereas
the Pods haven't been upgraded yet. It may also report a lower version
and no Pods are running any more with that version.
For other resources, we just check the Pods.

Parts of the Status is refactored into a common `DeploymentStatus` for
Kibana, ApmServer and EnterpriseSearch. Elasticsearch and Beat have
different requirements which cannot fit this common place.

This is a pre-req for strict stack version upgrade ordering
(https://github.com/elastic/cloud-on-k8s/issues/2600).