Add metadata for transform node role  (#3411)
Elasticsearch 7.7.0 introduces the transform node role.
-   add this new node type to the pod metadata when running a version of Elasticsearch that supports transform.
-   adjust test cases and default values.

The complication here is that we now have version specific default configurations. I chose to just toggle the transform value off in the default config to be able to keep using just one config struct. Technically the transform value does not even exist prior to 7.7.0. The default behaviour differs depending on whether the node is a data node or not. On nodes with node.data: false node.transform is also false unless specified otherwise explicitly.


Co-authored-by: Thibault Richard <thbkrkr@users.noreply.github.com>