Annotate the version of referenced resources in associations
As part of the association controller, retrieve the currently running
version of a referenced resource (from `status.version`),
and store it in an annotation of the resource holding the annotation.

Example: in the APM resource, set an annotation with the currently
running Elasticsearch version, and another annotation with the currently
running Kibana version.

We reuse the existing associationConf annotation here, for simplicity.

The version may be empty (`""`) if unknown (example: resource still in
the process of being created).