Delay version upgrades based on associated resources version
As part of the reconciliation of each resorce (except Elasticsearch):
inspect the version of referenced resources. If lower, do not move on
with the version upgrade yet, but requeue.

Example: in the reconciliation of Kibana 7.8.0, if the referenced
Elasticsearch resource is running version 7.7.0, requeue early before
doing any Kibana upgrade. Once Elasticsearch runs version 7.8.0, we can
move on with Kibana reconciliation.

Note that we don't try to be smart here in applying partial changes (eg.
configuration, resources, etc.) before the version upgrade is allowed.