Watch Pods in all controllers (#3534)
We did not enqueue reconciliations on Pods changes in Kibana, APMServer
and Beats controllers. We did on Elasticsearch and EnterpriseSearch
controllers.
When controllers rely on information retrieved from Pods as
part of the reconciliation, not watching Pods can lead to missing some
events.
This is especially true for the `status.version` field in all resources.
Note this particular case could have been worked out differently, by
checking the replicas sets for example. I think the implementation would
be more complicated, and we already watch all Pods due to the
Elasticsearch and EnterpriseSearch controllers anyway.

This commit sets up a single helper function called from all
controllers.