Use a kustomize patch to ensure a top-level additionalPrinterColumns
Having additionalPrinterColumns set at the `versions[]` array level in
the CRD is not compatible with k8s 1.11.
This commit patches the generated CRD to move the
additionalPrinterColumns of the v1 version at the top-level.

If this new CRD is used with an operator running an older version (which
is unlikely to happen), or during the first seconds before a
reconciliation happens with ECK 1.3, then the `version` field appears
empty in `kubectl get elastic`.