Fix makeObjectSteps tests name (#3556)
We printed test names such as `Create /`, which does not make sense.
I think the initial idea was to print `Create <namespace>/<name>`, but a
bug in the code printed <namespace>/<namespace> instead.

However it turns out some resources created there don't have a namespace
(example: ClusterRoleBinding). I think it makes more sense to print the
resource kind and the resource name, and ignore the namespace.

Also remove the `/` which is already used to render a test hierarchy in
the test output.

So in the end this commit does what's necessary so we get something like
`Create ServiceAccount filebeat-dvqx` instead of `Create /`.