Enforce a strict upgrade order between stack components (#3537)
Ensure individual stack resources can only have their version upgraded if referenced resources are also upgraded.
Implicitly, this implements the following upgrade order:

- Elasticsearch > Kibana > Beats, APM
- Elasticsearch > EnterpriseSearch

Example: if a user upgrades both Elasticsearch and Kibana from 7.7.0 to 7.8.0, the Kibana upgrade is delayed until Elasticsearch runs 7.8.0.

This restriction only applies to major and minor versions. Patch versions are ignored: Kibana 7.8.1 can be deployed alongside Elasticsearch 7.8.0.

In order to do so, the association controller adds a version field to the existing associationConf annotation, in the resource holding associations.
Example: the ApmServer resource that specifies a reference to Kibana and Elasticsearch will have the (lowest) currently running Kibana version and currently running Elasticsearch version in its annotations, set by both association controllers.

Each resource controller can then inspect its own association configurations annotations, and requeue in case the referenced resource is running an older version. Or if its version is not reported yet (resource in the process of being created): this slows down the initial resource creation slightly (Kibana is not created before the first Elasticsearch Pod is created), but I think that's short enough to not bother much.

We do not try to be smart in decoupling version upgrades from other upgrades (for example: decide to apply a config change but not a version change yet even though the spec specifies both): this would be tricky to implement.

**Breaking change**: we used to allow several associated resources to coexist with different versions, and apply any spec change immediately. We now enforce a strict version upgrade ordering that delays any spec changes until the resource version allows it. For example: when updating both Elasticsearch and Kibana from 7.7.0 to 7.8.0, any change to the Kibana spec will be delayed until Elasticsearch has been successfully updated to 7.8.0. Only major.minor in the version is considered by this.