Update owner references in reconcile resource util function (#3574)
We previously set owner references only on creation as they usually don't change. With the association model in ECK ownership can change though. The referenced resource can change and the controller e.g. the Elasticsearch controller for a Kibana -> Elasticsearch relationship relies on the owner referrence in its watch on secrets.

This commit takes this into account by resetting the owner reference on updates.