Reflect currently running version from status.version in additionalPrinterColumns (#3549)
* Reflect version from status.version in additionalPrinterColumns

When running for example `kubectl get elastic`, we get the version
displayed from `spec.version`. It represents the desired version but not
the version currently running.

Now that we have `status.version` available to represent the version
currently running, let's use it!

* Use a kustomize patch to ensure a top-level additionalPrinterColumns

Having additionalPrinterColumns set at the `versions[]` array level in
the CRD is not compatible with k8s 1.11.
This commit patches the generated CRD to move the
additionalPrinterColumns of the v1 version at the top-level.

If this new CRD is used with an operator running an older version (which
is unlikely to happen), or during the first seconds before a
reconciliation happens with ECK 1.3, then the `version` field appears
empty in `kubectl get elastic`.