Add a bit more context in data integrity check errors (#3679)
In order to investigate https://github.com/elastic/cloud-on-k8s/issues/3676,
this commit adds a bit more context to the api errors returned by Elasticsearch
in DataIntegrityCheck.Init().

The goal is to understand if there's a common pattern in these errors (eg. always
returned when the first document is indexed).