Retry getting the default storage class in E2E tests
On AKS clusters, the default storage class regularly takes some time to
be created. On one of my local tests, it got created after 4min.

This commit sets up some retries (20 with 30sec between each), so we
don't fail the AKS E2E tests if the default storage class is not there (yet).