Avoid panic on Kibana telemetry API call (#3736)
In 7.9, a change in the Kibana's startup behaviour can lead to have `/api/status` to return green 
while Kibana is not completely ready. Then, the Kibana telemetry API call may return an empty 
response and causes a panic.
This commit fixes this by retrying the call until the slice is not empty.