Set a temporary ownerref on Pods while the StatefulSet is being recreated
If the Elasticsearch resource is removed while the StatefulSet is being
recreated, orphan Pods will stay around forever.
To avoid that situation, we can set append a temporary owner ref to the Pods
before deleting their StatefulSet, so Pods don't stay orphans if Elasticsearch
is deleted.
We can then remove the temporary owner ref that was set after
the StatefulSet is recreated.