Support Elasticsearch volumes expansion (#3752)
Add support for resizing Elasticsearch volumes, by simply editing the storage requests in the volumeClaimTemplates section of the spec.

## Constraints

Volume size can only be increased, not decreased (k8s limitation, not likely to change anytime soon)
Only possible if the storage class specifies allowVolumeExpansion: true
Hot filesystem resize without re-creating the Pod is possible in many cases (GCE-PD, AWS-EBS, Azure, TopoLVM, etc.). If the storage driver does not allow it, it is necessary to manually delete the Pods. ECK does not do it.


## Implementation overview

Resizing PVCs is not supported by the StatefulSet controller at this time. The volume section of the StatefulSet spec is immutable.
However, it is possible to work around the StatefulSet limitations this way:

Update the storage requests spec of all existing PVCs: they are immediately resized by the storage driver, if inline expansion is supported. Otherwise Pods need to be recreated.
We now have resized volumes, but the StatefulSet still specifies the old volume size. Any new replica will end up with a smaller PVC, that in turn needs to be resized.
Delete the StatefulSet (but not the Pods it owns), then recreate it with the new volume spec.
Address the case where Pods may become orphans if something happens while the StatefulSet is getting recreated, by ensuring we always recreate deleted StatefulSets, and ensure we set a temporary ownerRef on the Pods while they are not owned by the deleted StatefulSet anymore.





* Remove the PVC validation webhook

We'll reintroduce it later, in a form that allows modifying
PVC storage reqs.

* RBAC access to read storage classes

* Move GetClaim to a common place for reuse

* Rename pvc.go to pvc_gc.go

* Resize PVCs whose storage req has changed

* Delete StatefulSet for recreation on volume expansion

* Pass the replicas count instead of the entire sset to LimitMasterNodesCreation

* Take into account temporarily deleted ssets when adjusting replicas

* Plug the volume expansion logic to the upscale phase

* Remove deprecated rolling upgrade partition strategy from upscale tests

* Garbage collect PVCs after having dealt with temporarily deleted ssets

* Use the highest existing ordinal instead of the pods count

...when setting the initial replicas values for statefulset re-creation.
Otherwise, if we're unlucky and one Pod is missing for some reasons (eg. pod-0),
using the Pod count could lead to removing the Pods with highest ordinals.

* Add missing license headers

* Improvements from PR review

* Revert "RBAC access to read storage classes"

This reverts commit 41e750f8bce3da4eaa1883aa983f19db4f17d0a1.

* Remove storage class checks for rbac concerns

* Revert "Remove the PVC validation webhook"

This reverts commit 000e05ab75a0a10cb5aaf6b1af00f12f347b0888.

* Ignore storage requests changes in the pvc validation webhook

* Reword the pvc modification error msg

* Store statefulsets to recreate in an annotation

Before this commit, it was possible to end up with orphan Pods in case
the Elasticsearch spec changed after a StatefulSet was just deleted.
This commit ensures deleted StatefulSets are always recreated, by storing
their spec in an annotation of the ES resource before doing the actual deletion.

* Remove the special upscale state fix not needed anymore

* Improvements from PR review

* Set a temporary ownerref on Pods while the StatefulSet is being recreated

If the Elasticsearch resource is removed while the StatefulSet is being
recreated, orphan Pods will stay around forever.
To avoid that situation, we can set append a temporary owner ref to the Pods
before deleting their StatefulSet, so Pods don't stay orphans if Elasticsearch
is deleted.
We can then remove the temporary owner ref that was set after
the StatefulSet is recreated.

* Fix unit tests

* Fix unit tests