Document volume expansion and volume claim immutability
This commit modifies the existing documentation in order to account
for the volume expansion feature introduced in https://github.com/elastic/cloud-on-k8s/pull/3752.

With those changes, it is now possible to resize storage requests if the storage size
allows it.
Every other change is forbidden.

The doc also mentions how Pods may need to be restarted if the volume driver
does not support online expansion.

I added lower-level titles to better structure the volume claim templates page.