Validate the storage class allows volume expansion before resizing volumes (#3795)
Before moving on with resizing PVCs, inspect the storage class referenced by claims to resize: if it does not allow volume expansion, error out. This is done both in the reconciliation loop, and in a validation webhook.

:: Storage class validation

Retrieve the existing StatefulSet matching the expected NodeSet. If a storage increase is requested: retrieve the corresponding storage class, and ensure it has AllowVolumeExpansion: true. In case no storage class is specified, we retrieve the default storage class, as indicated by its annotations.

This is done as part of the reconciliation, before attempting to resize volumes, but also in the validation webhook.
In the webhook case, it's important to realize we compare volumes in the expected Elasticsearch resource with volumes in the existing StatefulSets from the apiserver, and not with volumes in the "old" Elasticsearch resource. The reason is we want to generally reject storage decrease, except in the following case:

user increases the storage on the ES resource (maybe with the webhook disabled)
it turns out the StatefulSet cannot be updated with that new storage: user notices an error in the operator logs
user sets back the storage size to its old value: we want this storage decrease to succeed!
It also has the benefit of making the storage size check exactly the same in the webhook vs. the reconciliation loop (both check the expected volumes against existing SatefulSet volumes).

The validation code lives in validation/validations.go.

:: Webhook refactoring

In order to retrieve StatefulSets and storage classes, we need the validating webhook function to use a k8s client. This is not made easy by the controller-runtime webhook framework which cannot possibly hold additional parameters. I decided to reimplement the Elasticsearch webhook using lower-level primitives (the ones actually used by the controller-runtime webhook framework behind the scenes).

Keeping the refactored webhook in apis/elasticsearch/v1 creates a weird circular dependency because it relies on our k8s client wrapper, which itself relies on apis/elasticsearch/v1 to inject the right scheme. Also it feels weird to have some complicated logic involving checking resources in the apiserver in the apis package.
For that reason I moved all the webhook and validation code of Elasticsearch in controller/elasticsearch/validation. Functions there are called both by the controller itself (to perform validations in the reconciliation loop) and by the manager main.go (to setup the webhook).

:: Putting storage class checks behind a flag

Storage classes are cluster-wide resources. In some situations the user may not be able to give RBAC access to those resources. That's why we have the restricted helm chart profile to deploy the operator.
I added an additional flag to enable the storage classes validation (defaults to true), and tweaked the helm chart to:

set the flag to false in the restricted profile
not create the storage class rbac rule if no clusterScopeResources should be created