Allow watching cluster-scoped resources when using the multi-namespace cache
When managing resources in a single namespace, we rely on the controller-runtime
opts.Namespace feature, which implicitly allows watching cluster-scoped resources.

When managing resources in more than one namespace, we rely on opts.NewCache,
which takes a function to setup a multi-namespace cache. This cache does not
allows watching cluster-scoped resources by default. Which leads to error
when it tries to get storage classes:

```
cannot retrieve storage class: unable to get: /standard because of unknown namespace for the cache
```

This commit fixes it by ensuring we always append the empty namespace ("") to the
multi-namespace cache, so we can retrieve cluster-scoped storage classes if allowed
by the --validate-storage-class flag.