Don't allow downscales if some shards are unassigned (#3883) (#3887)
* Don't allow downscales if some shards are unassigned

In some conditions, for example when a Pod gets killed/restarted right
before a downscale happens, the shards on that Pod is reported as UNASSIGNED.
At this point it is dangerous to definitely remove the Pod (downscale the cluster)
since we can't know for sure the Pod to remove isn't supposed to hold any of the
unassigned shards.

To avoid that situation, this commit disallows any downscale to happen if
some of the shards don't have a node assigned to them (regardless of their status -
unassigned or not). The logic to allow a node to be downscaled is rather simple:
- all shards must have a node assigned to them
- the pod to remove must not have a shard assigned to it

This is a rather conservative/safe approach that could be optimized in the future.

I tried implementing an e2e test for this, but it's a bit tricky to setup the right way
so it consistently fails without this commit. I'm considering the unit test is good enough.

* Improve comments

Co-authored-by: Anya Sabo <anya@sabolicio.us>

Co-authored-by: Anya Sabo <anya@sabolicio.us>

Co-authored-by: Anya Sabo <anya@sabolicio.us>