Include secret token in APM Server config checksum (#3858)
The secret token is injected via an environment variable into the config file which is why changes to the token where not taken into account when building the config checksum. By adding the token value explicitly to the checksum APM Server pods will be rotated when the secret token changes.