Sort builders by test priority in Beat recipes E2E tests
In those tests we processed the builders in the order they are parsed
(Beat first, then Elasticsearch & Kibana).

In some cases a Beats test fails because Beats is waiting for Elasticsearch
to be available. The test failure reports a Beats problem, but the real
problem is on Elasticsearch side, which is misleading.

This fixes it by reordering builders by priority after parsing them
from recipe files.