Better error message on E2E CheckESNodesTopology failures
There are 2 reasons why  some nodes reported from the Elasticsearch cluster cannot be
matched to expected nodes in the spec:

1. Those nodes are being downscaled (not in the spec anymore), should be removed soon
2. Those nodes do not match the spec (different config, name, memory limit)

In the 2nd case, the error we report is rather generic. We print the expected topology
that hasn't been matched to the actual nodes: `expected elements missing from cluster`.

This commit adds an additional error message earlier in the process to report more details
about the actual node that cannot be matched to the spec (name, roles, memory limit).

Would have helped troubleshooting https://github.com/elastic/cloud-on-k8s/issues/3917.