Use new node_names query param for voting exclusions as of 7.8.0
The new query parameter tolerates non-existing nodes and covers therefore the case where we try to remove a master node that has never successfully joined the cluster. This prevents and orchestration dead lock where the operator was previously unable to make progress due to an API error returned by Elasticsearch.