Use new node_names query param for voting exclusions as of 7.8.0 (#3950) (#3954)
The new query parameter tolerates non-existing node names and covers therefore the case where we try to remove a master node that has never successfully joined the cluster. This prevents an orchestration deadlock where the operator was previously unable to make progress due to an API error returned by Elasticsearch.