Don't set an ownerRef on secrets users are susceptible to copy around (#3992)
* Don't set an ownerRef on secrets users are susceptible to copy around

A k8s bug (https://github.com/kubernetes/kubernetes/issues/65200) may cause the
k8s garbage collection to delete undesired resources in case users manually
copy an operator-managed secret to another namespace.

To avoid that situation, this commit ensures no ownerRef is set on a subset of
managed secrets users are susceptible to copy around:
- the elastic user password secret
- elasticsearch public transport certs
- elasticsearch, kibana, enterprise search, apm server public http certs

Existing ownerReferences set with earlier ECK versions will be removed when
reconciled.

Since they do not have an ownerRef anymore, those secrets are not automatically
deleted when the Elasticsearch resource is deleted. To work around that situation,
the secret reconciliation logic adds an additional set of labels to the reconciled
secrets that don't have an ownerRef specified. These labels reference the "soft"
owner ("soft" as in handled through some custom code and not through k8s builtin
garbage collection logic).
Once a controller receives a deletion event for the resource it manages, it will
automatically remove the soft-owned secrets.
This is done as best-effort. Secrets will remain orphan if:
- the operator is not running when the owner is deleted
- an error happens while deleting the soft-owned secrets

* Error out the reconciliation on garbage collection errors

* Improvements from PR review

* Label soft-owned secrets with the owner Kind

* Reuse Kind constants instead of hardcoded string

* Fix existing unit tests

* Garbage collect orphan soft-owned secrets on operator startup

* Fix linter warnings

* Add unit tests

* Cover soft owned secrets gc in e2e tests

* Fix linter warning (typo)

* Improvements from PR review

* Trigger reconciliations on soft-owned secrets events

* Fix linter stuff